msm: kgsl: Do not delete command batch unless it's retired
Do not remove command batch from the dispatcher list if it's timestamp
has not retired. Even if the context to which the command belongs
has been detached. If the command is in the dispatcher queue it means
it's in the ringbuffer and we should wait for it to complete the normal
way. Also make sure that we use the correct retired timestamp even if
the context was detached because of a userpace context destroy call.

Change-Id: If3a9a562180b924492ed95f208b5e3d469abdfba
Signed-off-by: Shubhraprakash Das <sadas@codeaurora.org>
Signed-off-by: Carter Cooper <ccooper@codeaurora.org>