msm: kgsl: Define new fault type for IOMMU fault
When IOMMU fault happens then set a fault bit defined
for it to handle it as a soft fault case instead of a hard
fault. This is required because a different fault can trigger
due to a page fault since we stall the gpu. If that happens the
other fault handler will run and handle the fault after which
we may end up in fault handler again trying to handle a false
IOMMU fault this time around. Since, its a false fault now we do
not want the command to be skipped which a hard fault enforces
and if we are in a true IOMMU fault then the command will be
skipped anyway since we check of a pagefault specifically in
the context.

Change-Id: Ib8e02190aeee911f9ce892b683e756528e5f8865
Signed-off-by: Shubhraprakash Das <sadas@codeaurora.org>
Signed-off-by: Carter Cooper <ccooper@codeaurora.org>

Conflicts:

	drivers/gpu/msm/kgsl_iommu.c